GUI support for VF interface type
This commit provides support for a new interface type: 'VF'.

A VF (virtual function) interface is a sub-interface of an existing
pci-sriov class interface.  It can have a different (less than)
number of VFs as the parent interface as well as a different
virtual function driver.  In addition, the VF interface can be
assigned to a different data network as the parent SR-IOV interface.

The purpose of the VF interface is to enable a user to split the
virtual functions of a parent device between the parent and child
so that each can be assigned a different VF driver. This may be
beneficial to a user that has a mix of (DPDK) accelerated
containers and containers that just depend on a kernel
netdevice.

Depends-On: https://review.opendev.org/#/c/696372
Story: 2006842
Task: 37422
Change-Id: I2a5be1028c1a6b8fd39ffd8770e2e074069aaa58
Signed-off-by: Steven Webster <steven.webster@windriver.com>