Align OS_AUTH_URL in admin-openrc.sh
The admin-openrc.sh can be downloaded from Horizon. This file can be
used for authentication as sysadmin in system controllers. However,
as the keystone public endpoint of region SystemController differs
from the RegionOne. If a user download an RC file of SystemController
and use the OS_AUTH_URL to authenticate, a HTTP 401 Error(The request
you have made requires authentication) will be produced.

In the upstream project, the OS_AUTH_URL is got according to the
"Central Cloud Region" and shown in the view. This commit overwrites
the openrc template, aligns the OS_AUTH_URL in the admin-openrc.sh of
region SystemController with the RegionOne by port replacement. As it
is a specific usage for starlingx rather than a generic usage, it will
not go to the Horizon project.

Test:
1. Choose the "Central Cloud Region" in Horizon as "SystemController"
2. Download the admin-openrc.sh via API Access -> Download OpenStack
RC File
3. Check the OS_AUTH_URL is pointing to then keystone pulic endpoint
of RegionOne
4. Check the keystone pulic endpoint is still correct in the web page

Change-Id: I1f43f79364f5cc7bff382c1ae90a7f8f801abedb
Closes-Bug: 1892090
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>