Horizon GUI disconnects during subcloud actions
Currently after an unlock is executed and the horizon page is refreshed,
the subcloud system list is retrieved. If there is only one subcloud
controller an exception is thrown since there is no connectivity.

This change catches the exception and redirects to RegionOne where
central system controllers will be listed

Closes-Bug: 1939313
Signed-off-by: Jose Infanzon <jose.infanzon@windriver.com>
Change-Id: Ie8b5a785f0bdde7fb0481610d0f9a17b2b978a29