Warning message: outdated strategy
Confirmation before applying strategy will be needed.
If the created strategy is older than 60 minutes,
a warning message will be display on this popup.

The user can apply the strategy or verify if it is still valid.

Strategy Steps (valid subclouds for this strategy) is defined
when strategy is created. If subclouds status or Subcloud Groups
are modified before applying, there might be some inconsistences.
Therefore, after some time, a warning message will be display and
the user must decide whether or not to create a new strategy.

Story: 2008931
Task: 43215

Signed-off-by: Fernando Theirs <Fernando.Theirs@windriver.com>
Change-Id: I975aa6e7a1102467ae026e6539d8abdd88db147f