Fix missing "year-month-day" in log text of login
The timestamp of the login log text present in file
"/var/log/horizon.log" is missing the "year-month-day" portion. This may
cause problems when filtering logs by date.

Test Plan:

PASS: Successfully deploy an AIO-SX and verify that the Horizon page is
working correctly.
PASS: In the deployed AIO-SX, execute the login in Horizon page and
check that the timestamp of the login log text, present in
"/var/log/horizon.log", now contains the "year-month-day" part.

Story: 2009824
Task: 44651

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I9da29f8df3576506f4ed4dc82641a55ca7dfefef