Fix for 'Apply to' and 'Force' fields behavior
On 'Create Strategy' form the following fields are not usable when
selecting a subcloud:
* subcloud apply type
* maximum parallel subclouds
Now they are hidden in such case.

Input element for subcloud name was changed to a dropdown list.
If 'All subclouds' is selected, the strategy is applied to all
subclouds and the fields mentioned are hidden. If any other
subloud name is selected, such fields are shown.

It was created another 'force' element for kubernetes strategy
which shows a help message that is consistent with the expected
behavior. The original element is left for 'upgrade' strategy only.

Test Plan
PASS:
Build package
Install package on DC with at least one subcloud
Open gui on a browser and navigate to Distributed Cloud Admin ->
Orchestration -> Orchestration Stragegy -> Create Strategy:
* Check 'force' parameter is hidden for 'Patch' and 'Firmware' strategy
  types.
* Check 'force' help text describes actual behavior.
* Check 'subcloud' field is a dropdown menu and shows available
  subclouds.
* Check 'subcloud apply type' and 'maximum parallel subclouds' fields
  are not shown when a single subcloud is selected in the dropdown
  menu.

Closes-bug: 1973210
Closes-bug: 1973805

Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>
Change-Id: I909bb548f9b6454f6e556109165565394ec63114