Show 'error_description' field in Subcloud detail
As part of improving the error reporting on subcloud
deployment and upgrading, this change adds the 'error_description'
field in the list of attributes shown in Subcloud details.

Test-Plan:
PASS: Build and install rpm.
PASS: Deploy an ISO with these changes and see the new field added in
the subcloud details.
PASS: Upgrade from previous version to this version and verify the
changes are added as expected.

Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/857715

Story: 2010271
Task:  46297

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: Ib19713d27ff9195fc6183bf13daa4e9f31f14891