Change pages permissions for reader users
This change allows users with reader role to view the same pages of
users with admin role. The actions are limited by the various APIs
called but not by horizon itself.

Test Plan:

PASS: Successfully deploy an AIO-SX using a Debian image with this
change present and create user "readeruser" with reader role. Log in the
Horizon interface using "readeruser" user and check that the pages
"Software Management", "Host Inventory", "Data Networks", "Data Network
Topology", "Storage Overview" and "System Configuration" under "Admin"
-> "Platform" and the pages "Active Alarms", "Events" and "Events
Suppression" under "Admin" -> "Fault Management" are listed and are
accessible with no errors.
PASS: In the deployed AIO-SX, log in as "readeruser", go to "Admin" ->
"Platform" -> "System Configuration", click in "Edit System" in
"Systems" tab, try to change the system description to "Test Description
001" and check that it fails. Do the same as user "admin" and check that
it succeeds.
PASS: Successfully deploy a DC with 1 subcloud using a Debian image with
this change present and create user "readeruser" with reader role. Log
in the Horizon interface using "readeruser" user and check that the page
"System Configuration" under "Admin" -> "Platform" and the pages "Cloud
Overview", "Software Management" and "Orchestration" under "Distributed
Cloud Admin" are listed and are accessible with no errors.
PASS: In the deployed DC, log in as "readeruser", go to "Admin" ->
"Platform" -> "System Configuration", click in "Edit System" in
"Systems" tab, try to change the system description to "Test Description
001" and check that it fails. Do the same as user "admin" and check that
it succeeds.

Story: 2010149
Task: 46568

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: Ic94ae516afb21664d4cd20bd29080d3cde323083