Remove unnecessary button visibility logic
Current logic compares number of interface and number of
interface type for "Create Interface" button visibility.

Current interface type defines 4 types, so that if more than
4 interfaces are configured in any type, "Create Interface"
will be invisible, which should not.

This fix is to remove this logic, so that user can add
interfaces through gui.

Closes-bug: 1998638

Test Plan:
PASS: "Create Interface" button is visible
      after adding 8 interfaces and we can
      add more than 8 interfaces.

Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>
Change-Id: I409a3483d6e0c4e0f53436d02d710b4dcafe2d20