Fix 'dictionary changed size during iteration'
The Python "RuntimeError: dictionary changed size during iteration"
has been seen on some points, e.g. when creating a strategy.
It occurs when the size of a dictionary is changed during iterating
over it.

To fix this error is needed to add a '.copy()' method to create a
shallow copy of the dictionary.

Test Plan:

PASS: Apply these changes and create a strategy. Verify the strategy
is being created successfully. Same test has been done for other
sections.

Closes-Bug: 1999069

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I3f6dc0e27547d6439a2d14e9ba8ee6dd4f68e8be