Add "Apparmor Module" on Edit Host modal
This change adds a dropdown menu in order to enable or disable
the apparmor module when updating a host.

By default, apparmor will be disabled on creation and this value can
be modified using the Edit Host button, or system host-update
command.

Test Plan:

PASS: Build and install starlingx-dashboard package. Verify that
the changes are applied.
PASS: Edit a locked host in order to change the apparmor value.
Verify that the change impacts when running system host-show command.
PASS: Edit a unlocked host by changing the apparmor value. Verify
that an error message is shown since the host has to be locked.
PASS: Verify that the correct apparmor value is shown in
"Host Detail" page.

Story: 2010310
Task: 47249

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I814896806479e141ce86961cc08b48da5600afed