Add new fields in the GUI to edit filesystem size for root var and log
The new filesystems - var, root and log - are not able to be modified
from the GUI, because the form does not show these filesystems.
This fix adds the new filesystems so they can be modified.

Closes-Bug: 2008810

Test Plan:

PASS: Modify the new filesystems from the GUI. If the size is
decreased, the appropriate error message will be displayed.
PASS: Modify the other filesystems from the GUI. If the size is
decreased, the appropriate error message will be displayed.

Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>
Change-Id: Idfe3b48762708fbed5f7d029445590667c1c43cc