Fix to-version field on K8S Upgrade Strategy
When creating a Kubernetes Strategy, the 'to-version' field
was being sent without the 'v' prefix. This was making that the
kubernetes pre-check fails.

This change fixes this issue. Also, the default version that is
selected when the user doesn't select an item on the 'to-version'
drop-down menu will be the systemcontroller active version.

Test Plan:
PASS: Create and apply a kubernetes upgrade strategy selecting the
desired 'to-version' option. Verify the pre-check is passed and
the strategy is completed.
PASS: Create and apply a kubernetes upgrade strategy using the default
option on the drop-down menu. Verify that the version that it being
sent to dcmanager is the systemcontroller's active version.

Closes-bug: 2018611

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: Ifb76950dfdc67bdda57cbaea20433f65ebcc8aa3