Horizon orchestration pop-up shows redundant error
When you try to create an orchestration Strategy and it failed, the
pop-up showed the same error message in the main and details section.

This change fixes this issue. In addition, has been fixed for the
Apply Strategy action too.

Test Plan:
PASS: Create an invalid strategy and verify the pop-up error message
is displayed properly for example:
Error: Strategy creation failed.
Details: The server could not comply with the request since its either malformed or otherwise incorrect. Invalid groud_id
PASS: Apply strategy and verify the pop-up error message
is displayed properly.

Closes-bug: 2018625

Change-Id: I805cc051a843af07f2dba1701aef8bc791b9450a
Signed-off-by: Rafael Moyano <rafael.moyano@windriver.com>