Fix OS_AUTH_URL included in OpenStack RC file
When the OpenStack RC file is exported from Horizon,
the URL in the OS_AUTH_URL variable does not include
the /v3 API version.

When the system is configured in secure mode (HTTPS
enabled), an error is generated during the login
process when trying to use any of the StarlingX
CLIs (platform, fm, sw-manager, etc).

This issue can be resolved by appending the API
version to the OS_AUTH_URL.

Test Plan:
    PASS: Log-in Horizon, download openrc.sh file.

    Using a remote CLI:
      PASS: source downloaded openrc.sh file.
      PASS: Test Platform CLI (ej: system host-list)
      PASS: Test FM CLI (ej: fm alarm-list)
      PASS: Test sw-manager CLI (ej: sw-manager patch-strategy show)

    PASS: Enable secure-mode (HTTPS with Self-Signed Certificate)
    PASS: Log-in Horizon, download openrc.sh file.

    Using a remote CLI:
      PASS: source downloaded openrc.sh file.
      PASS: Test Platform CLI (ej: system host-list)
      PASS: Test FM CLI (ej: fm alarm-list)
      PASS: Test sw-manager CLI (ej: sw-manager patch-strategy show)

Closes-Bug: 1998949

Change-Id: I5c678a6ad89f9d5e56356189962e8c09d3ffa6c5
Signed-off-by: Jorge Saffe <jorge.saffe@windriver.com>