Add the upload-only option to the Horizon upgrade orchestration UI
This change adds the new DC upgrade orchestration upload-only option to
the upgrade orchestration UI. A new 'Upload Only' checkbox was added to
the 'Create Strategy' form when the selected strategy type is 'Upgrade'.

In the 'Orchestration Strategy' tab, if the current strategy is a upgrade
strategy, a new field 'Upload Only' with its current value is now shown
so that the user can know if an existing strategy is using this option.

Test Plan:
1. PASS - Verify that the 'Upload Only' checkbox appears correctly
          inside of the create strategy form only when the selected
          strategy type is equal to 'Upgrade';
2. PASS - Create a upgrade strategy with and without the upload-only
          option and verify that it gets created successfully;
3. PASS - Verify that the interface shows the 'Upload Only' field and
          its current value on the 'Orchestration Strategy' page only
          when using the patch and upgrade strategy type;
4. PASS - Create and delete another orchestration type (e.g. Kubernetes
          upgrade) through Horizon and verify that it still works as
          expected;

Depends-On: https://review.opendev.org/c/starlingx/distcloud-client/+/894169

Story: 2010676
Task: 48771

Change-Id: Id20dd42939165947eb377965557e9618e3b6b7ea
Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>