Update to_version field to return all kube versions
In this commit, the to_version field was updated to return
all kube versions and allow the user to install older versions.

Test Case:
PASS - Using Horizon, check all k8s are available under
drop-down menu.
PASS - Using Horizon, create and apply kube orchestration
with --to-version selecting a version available and verify
that the kube was upgraded to the correct version.


Closes-bug: 2043514

Change-Id: I72d5edd33188935e706649c8561b0d6eb7b9b6c6
Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>