Unable to remove NIC from bond in Horizon
Unable to remove a lower interface from a mgmt type interface
on controller-0 using Horizon.

The values in the 'uses' field received from Horizon
requests are the UUIDs of the lower interfaces. The sysinv
code expects ifname.

The conversion was already handled in adding
an interface, but not when updating an interface.

Testing:
- Build successful
- Test removing any lower interface and adding it back
- Test removing lower interface related to the original mgmt mac
  -> verify that it does reselect the original mgmt mac

Closes-Bug: 2066400

Change-Id: Ifd51814f90cabee88c5457128ad6ddbdac2e975c
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>