Replace Subcloud Group text input field for a select field
The modal box for creating a Prestage Strategy
on the Subcloud Group is currently displaying
a text input field where the name of the subcloud
group must be typed.
This change replaces the text field with a select field
that lists all available subcloud groups..

TEST PLAN:
[PASS] Build, install.
[PASS] Navigate to Distributed Cloud Admin > Orchestration >
Create Strategy and and attempt to create a new strategy
with "Apply to: Subcloud Group" Verify that the Subcloud Group
field displays a select box with all available subcloud groups.

Closes-bug: 2096933

Change-Id: I18923ba6221e5a8531641fa6d628b62dc66e4a58
Signed-off-by: Italo Lemos <Italo.doRegoLemos@windriver.com>