Dynamically retrieve the region one name
This commit removes the hardcoded "RegionOne" region name and instead
retrieves the region name dynamically from the openrc file.

This change prepares for a future update where DC services will be
deployed on a standalone system that uses a UUID as the default region
name.

When selecting the region name on the region dropdown, it will still
appear as "RegionOne", but it'll use the dynamic region name
internally. This is to keep the user experience the same by not
showing the UUID in the selector.

Test Plan:
01. PASS - With the SystemController region selected, go to the Cloud
    Overview page and click on "Alarm & Event Details" on the Central
    Cloud region. Verify that it correctly switches to the RegionOne
    region.
02. PASS - Repeat the previous test but click on "Host Detail"
    instead.
03. PASS - With the RegionOne region selected, attempt to add and
    update a interface.
04. PASS - With the SystemController region selected, go to the Cloud
    Overview page, then click "Alarm & Event Details" on a subcloud
    row. Verify that the subcloud name appears on the dropdown on the
    top of the page.
05. PASS - On the region selection dropdown, selection RegionOne and
    verify that it correctly switches to the RegionOne region.

Story: 2011312
Task: 51917

Change-Id: Ia7921fe46089297e4edfc613ea92b6c03376b93e
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>