Fix create vf interface
This change fixes the creation of a new vf interface.
The GUI was preventing the creation of a vf-type
interface because the max_tx_rate field was always
being submitted, which caused an error on some ports.
The same command works when executed via CLI.
An adjustment was made to the submission of
the max_tx_rate field to avoid this issue.

[PASS] Build, install
[PASS] Go to Host Inventory > Host Details >
Interfaces tab > try to create a new vf interface
type with and with not a max tx rate field.

Closes-bug: 2111263

Change-Id: I3a4af746803d42258c56bdee1072749d87e4d767
Signed-off-by: idoregol <Italo.doRegoLemos@windriver.com>