Fix support for rook-ceph storage
Horizon was not displaying store functions when rook-ceph
was enabled. This changes adds support for rook-ceph in
Horizon and correctly reads data from a storage backend
based in rook-ceph.

TEST PLAN
[PASS] Build, install with rook-ceph configured.
[PASS] Navigate to Admin dashboard > Host Inventory >
Select a host in Hosts tab to see the detail page >
Go to Storage tab verify if Storage Functions table is
displayed.

Closes-bug: 2116276

Change-Id: Ic78acf1ab21176558a6e9fd0d43abb9b91270412
Signed-off-by: Italo Lemos <Italo.doRegoLemos@windriver.com>