Remove cert-file upload in K8s RCA update
Remove the cert file option for the kube rootca update form. This
change is in line with the removal of the customization for the
K8s RCA from bootstrap and system/sw-manager/dcmanager CLIs.

Test plan:
PASS: Deploy DC Systemcontroller + SX subcloud.
PASS: Verified that the cert file option is not available in the
      K8s RCA update form.

Story: 2011604
Task: 53489

Change-Id: Ic92ab63b579496c0391675e125e22df516d74211
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>